+++ relm4/src/shutdown/sender.rs
@@ -2,26 +2,22 @@
 // SPDX-License-Identifier: MIT or Apache-2.0
 
 use async_broadcast::Sender;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
 
 /// Sends shutdown signals to receivers.
 #[derive(Debug)]
 pub struct ShutdownSender {
-    pub(super) alive: Arc<AtomicBool>,
     pub(super) sender: Sender<()>,
 }
 
 impl ShutdownSender {
     /// Broadcasts a shutdown signal to listening receivers.
     pub fn shutdown(&self) {
-        drop(self.sender.broadcast(()));
+        drop(self.sender.try_broadcast(()));
     }
 }
 
 impl Drop for ShutdownSender {
     fn drop(&mut self) {
-        self.alive.store(false, Ordering::SeqCst);
         self.shutdown();
     }
 }